requirements.txt
Flask==3.0.0
Pillow==10.1.0
reportlab==4.0.7
Werkzeug==3.0.1

# Optional dependencies for advanced features
# Uncomment if needed:
# geopy==2.2.0    # For geolocation features
# piexif==1.1.3   # For EXIF metadata handling
# pikepdf         # For PDF manipulation
# fpdf2           # Alternative PDF generation